pyproject.toml
[tool.poetry]
name = "libshipkore"
version = "0.1.2"
description = ""
authors = ["Gagandeep Singh <gagan@innerkore.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
beautifulsoup4 = "^4.10.0"
python-dateutil = "^2.8.2"
click = "^8.0.3"
pydantic = "^1.9.0"
parsel = "^1.6.0"
class-registry = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
